requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# AI APIs
openai>=1.0.0
anthropic>=0.25.0

# Data generation
faker>=20.0.0
python-dateutil>=2.8.0

# Configuration and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# Utilities
pytz>=2023.3
